requirements.txt
pyyaml>=6.0
requests>=2.28.0
gitpython>=3.1.0
google-cloud-aiplatform>=1.25.0
google-genai>=1.9.0
python-dotenv>=1.0.0
Flask
requests
python-dotenv
gunicorn
Jinja2<3.1.0
GitPython>=3.1.0